+++ resources/META-INF/plugin.xml
@@ -34,7 +34,7 @@
   <actions>
     <!-- Add your actions here -->
     <group id="actionGroup">
-      <action id="FileChangeManager.contextMenuActions" class="contextMenuActions"
+      <action id="FileChangeManager.FeatureSelectionAction" class="FeatureSelectionAction"
               text="ActiveDocumentation Feature Selection"
               description="ActiveDocumentation feature selection"/>
       <separator/>
